log: Don't panic if log.Error() is called with a nil error
Summary: Calling log.Error() with a nil error is an easy mistake to make and is difficult to spot until it happens. At that point there's a panic and now you have two problems. If it's a problem that nil errors aren't reported, the "nil error logged" message can be tracked and alerted on.