grpc: Improve processing of grpc errors
Summary: If the error doesn't have a status, we return the error. 

If the status doesn't have a jettison error, we wrap it in a jettison error and then a grpc Error struct. 

If we do get a jettison error out of the details, then we just wrap it in the grpc Error struct and return that.